Refactor: make document unit info panel re-usable
So far this component was only used by for caselaw and thereby was
a little specialized. As the template itself was already quite generic,
it was as simple as forwarding this abstraction to the component
properties. This allows the parent component to define the heading as
well as the list of properties to render specifically for the use-case.

While the component's unit tests had to become more generic, they now
test the correct order and positioning of the property infos.

Note that this component will become target of the component library.