Feat(norms): add info panel for norms
Please note that the name of the used component has still a caselaw
specific background related to their domain entity name `DocumentUnit`.
Unfortunately does design call this component internally
`DocUnitHeader`. As that names brings some issues with it, it stays as
it for now until we found a well fitting alternative.