Fix: update frontend base image for CVE
There was a new bunch vulnerabilities reported for the frontend image in
relation to some SSL libraries. Most importantly to mention the high
vulnerability of CVE-2023-0286.
To fix this, a new version of the base image used for the frontend is
necessary to get the libraries with the security patches.