Use axe core plugin
RISDEV-1062
Replaced the ‘axe-playwright’ package with ‘@axe-core/playwright’, as this is recommended in the official playwright documentation. Also this package is able to exclude certain rules or certain elements from accessibility checks. For the categories page, we disabled the rule ‘duplicate-id-aria’. Some background: We make the assumption, that the field id in the configs always matches the domain model. Input field IDs need to have the same value like the model name, e.g. the court input fields in coreData and previousDecision both need to have the ID ‘court’, because the model is named so. This results in a ‘duplicate-id-aria’ violation, we need some decoupling here, so we are able to use different ids in the frontend. Also the ModelComponentRepeater creates components with the same id, we will need some kind of indexing in the ids here. That's why we disable the duplicate-id-aria rule for now.