Test(norms): use private Juris test files
The import of a norm use-case has changed in the way that it expects
a Juris file as input now. These files are legally required to remain
private and can not be part of the test data in this repository. But as
the import of a norm is part of the setup for each norm end-to-end test,
all of them had to be skipped.

The solution is now to automatically fetch the private data within the
test setup and store them locally in a temporarily folder. The folder
acts as some kind of cache but is also required to allow the upload of
the file via file selector in the end-to-end tests and for drag and drop
functionality.
The files get downloaded form the private repository that includes the
`ris-norms-juris-extractor` library that has a bunch of test data for
its integration tests. The download uses the already existing secrets
of the environment variables from this repository.

The used base norm had to change, as the old one is no more part of the
exemplary data sets. The one with the most metadata fields filled was
chosen.