Fix: add machine user secret to e2e workflow
The problem was that the end-to-end tests are written in a different
workflow that is sourced by the pipeline workflow. Thereby, the child
workflow has no access to the secrets of the parent. In result it was
not possible to download the Juris test files for the norm end-to-end
tests.