Test(caselaw): fix data transfer for file dragging
We have some end-to-end tests that test the drag-and-drop behavior for
file uploads. Therefore an artificial data transfer event must be
created as there is no actual file browser or so to drag from.

The creation of this object was not 100% correct and uploads actually
failed as the backend was reporting a wrong file type. But this never
got detected as the tests do not actually test something right now. You
only notice it if running the tests visually in debug mode.

This fixes the creation of the data transfer and exports it into
a re-usable function so also others tests (e.g. for norms) can make use
of it.