Feat: allow button component to be a link
Design has the need to have links which look like a button. This means
the whole styling and behavior remains stable, but the actual HTML
element has to switch according to the use-case.
To make this possible, the `TextButton` component is using a manual
render function instead of a template. Using the new property of `href`
is used to alternate between a button or an anchor link. The visual
result is indistinguishable.