Feat: remove default label from text button
For unknown reasons, the `TextButton` component still had a default
label set to `Speichern`. As this doesn't make to much sense, it has
been removed. The two locations where this was actually "used" have
their own label definition now.
This also simplifies how to create an icon only button. This can be seen
in the changed made to the kitchen sink view.