Feat(norms): extend API specification for files
The load norm operation also needs to provide the list of files
associated with the loaded norm. Therefore the endpoint got extended and
will be implemented soon.

Please note that it is currently "not possible" to add an example
SHA-256 hash value to the specification. It causes huge trouble with the
security scanner and seems to be unresolvable for now.