Feat(norms): add new route to export norm
This route enables the user to download the Zip file for the jDV.

Note that this page is not yet available via the navigation to the user.

Also note that there are no end-to-end test for this route yet, as the
backend is not yet fully implemented for this feature.