Update spotless tooling and apply formatting
With the new prettier-plugin-sql version, the SQL files will be
formatted in a different way.

This causes the Flyway migration files to have different checksums
and the migration would fail.
Due to this, we need to first call `flyway.repair` to fix the checksum
across the project.

In an upcoming push, the `repair` command will be removed again.