Feat(norms): add metadatum entity to a norm
In regards of our strategy, we aim for a "new" structure of the
metadata. The first step is the new entity of a Metadatum. A norm can
now be associated with some metadata. For a starting point, only the
metadatum type of keywords is supported.

Note that this isn't used yet by any port and is also not yet supported
by the persistence layer.