Feat(norms): add specification concept to domain
We decided to use the concept of specifications from domain driven
design in our domain package. The currently primary use we focus on is
in-memory validation of data structures (entities and object values).

This simply adds the scaffold with the interface and chain operator
utilities. This is so simple and small, that no library or anything is
worth to be introduced.