Chore(norms): improve reactive database access
The primary focus was to improve the parallel access to the database
where necessary. For example it is not necessary to wait until all files
references for a norm have been saved before the articles can be saved.
This is in preparation for having very soon even more tables that can be
accessed in parallel.

Additional a couple of our reactive chains have been improved for
a better flow and readability.