Feat(norms): read and write metadata to database
Extends the respective output port implementations in the database
adapter to also save and read to and from the metadata table.

Includes the simplification to only have a single mapping function to
a norm entity. This requires to explicitly set certain parameters to
empty lists, which makes it more explicit.