Feat(norms): update metadata table on edits
Extend the implementation of the edit norm output port for the database
adapter to also update the metadata. So far, on editing, only the norms
table gets written with the changed data. The rest remains untouched.
This has changed, since editing now also includes the metadata which is
stored in a different table.
As metadata entries are not yet identifiable within the domain, there is
no mechanism to update a single row in the metadata table yet.
Therefore, the current mechanism first deletes all rows associated with
the edited norm and then (re-)writes all (new) metadata. This has been
agreed to be an intermediate solution.