Test(norms): fix flaky integration tests
The integration tests of the database have the issue that loading of
list values from other tables result into a random order. For the latest
changes this means that the order of the metadata is not stable. This
causes issues in tests which assert the equality of two objects.

This problem is already known and there have been exceptions for them in
the assertion function. This got improved by an approach that makes this
more generic. This should reduce the struggle and work more reliably
across changes.
Unfortunately does this not yet work for file references right now. This
was ignored yet, and bubbled up while working on the generic approach.
For the moment they are excluded.