Feat(norms): use metadata for juris converter
This change is still backwards compatible. This means that the keywords
are read from the metadata plus the old property. And the other way
around, extracted keywords get converted to metadata and put into the
old property.
Once the whole end-to-end scenario has made the change to the breaking
change, we finally can remove the old property.