Feat(norms): prune norms frame keywords property
After the finished transition of having keywords stored as metadata,
this property of the norm entity can be removed. It has been kept for
backwards compatibility. Now, all parts of the system have made the
transition.
This removal also includes the drop of the associated database column.
It has been decided in the team that isn't worth to migrate the old
data for this very specific use-case.