Fix(norms): fix flaky norms comparison in tests
The loading of list data from the database is unsorted. This means if
a norm gets loaded from the database, the order of the articles, file
references, paragraphs and metadata can be sorted slightly differently
every time. This is due to the reactive access.
Anyway, even if they would be strictly sorted based on any order, the
order of the same data as created (free by hand) can differ from the
sorting of the database.
In result, the comparison of two norm instance with exactly the same
data, but different orders, leads to (flaky) failing tests.

To fix this problem, a quite generic assertion functions was written
that uses reflection and recognized collections to apply a set on them
before comparison. Unfortunately it is pretty hard to do this on
a nested level where paragraphs within the articles are sorted
differently.
To have a solid solution for now, a function sorts all known list data
of a norm first and then compares them. The huge disadvantage is that
this is not generic and changes to the data structure can make it
ineffective again.