Feat(norms): add new data set summary component
This new component is indented to be very generic and be used other
a variation of many special cases. Though it still represents a shared
concept and visual part. It is intended to be used by both streams and
be very adaptive. Please read the test cases for some more details.

It is planned (and already locally experimented) to transform this to
be a type generic component. This allows to much higher type safety
and to better represent the actual logic of the component.

The application of this component is still in an experimental
development stage. Therefore it remains unused for now.