Feat(frontend): update vue-tsc for generic support
The Vue TypeScript compiler package supports in more recent versions
generic Vue components. This is a very powerful feature we are highly
interested in. The latest stable version is `1.2.0`. Though the patch
version `1.2.1` was necessary to overcome some issues. The next version
`1.3.8` is still too experimental, though it would work fine for us.

Updating the compiler came with two issues. First of all this "finally"
discovered our issue with the usage of `aria-label` as custom property
for many input components. Many colleagues have realized this issue in
their development environments earlier. After all that isn't a technical
issue, but requires to unmark this as a HTML exclusive property. The
default for `htmlAttributes` is `['aria-*']`. Therefore an empty list
`[]`, fixes this.
The second issue was with some changes to default slot properties. For
unknown reasons, this requires also the usage of template elements as
for named slots. An issue (vuejs/language-tools#2554) has been raised.