Feat(norms): use generic type for data set summary
This is the first application of a generic type for a Vue component in
the frontend. This allows us in this case to ensure that the passed data
property aligns with the related summarizer function. But also the data
property itself gets verified for consistency. The summary component
itself remains independent of the actual type, but enforces these
boundaries. This should also make the component more easy to understand
as it is more explicit. Mismatches will now be detected early during
development due to wrong typing. This can be already seen/recognized in
the tests.

As generics in Vue SFCs is still a quite new feature, there are some
little drawbacks which aren't fixed yet. Or rather the fixes are not
released yet. This relates to the property validation and the property
deep typing. Workarounds have been established with the respective
comments to resolve them later on. Nothing changes the overall code
structure and will be easy to be just deleted.