Feat: add expandable data set component
This component is intended to be re-used by the new design for all the
complex metadata sections of norms. It has a small feature set, but is
quite flexible.

There are some issues with the transition of the underlying
`ExpandableContent` component which become more visible with the new
intend usage. We might have to fix that soon.