Feat: adapt hiding mechanism for expandable
The `ExpandableContent` component now uses `v-if` instead of `v-show`.
This causes an unloading of the inner component instead of keeping it
"cached". This has become somewhat required for some new use-cases.

Unfortunately it is not possible for component slots to use a
`KeepAlive`. That would allow the usage of special hooks that could
simulate an unloading.