Feat: add new editable list component
This is basically the successor of the `ComponentModelRepeater` with new
design and updated behavior.

Something that is not yet properly working is the usage of the return
key to finish editing a list entry. Since our input components are still
lazy, the value of the last input will be missing and empty. That will
be automatically fixed soon.