Chore: remove obsolete nested button component
This component was a copy-paste of the original `TextButton` with minor
changes. But it is only used once at a single location for the
`EditableList`. As it has not yet proved to be a standalone concept by
design, it got inlined into the component itself. Thereby, this
component can be removed.