Feat(norms): improve lead input group component
This removes the usage of the `InputGroup` to overcome some
styling/spacing issues which are more customized in the new design.

There is some "ugly" CSS fix for the border of the left input when
hovered/focused. We need to iterate over it once any designer is healthy
again and can give feedback.