Avoid duplicate ids in frontend
RISDEV-1092
We make the assumption, that the field id in the frontend always matches the domain model in order to communicate with the backend. This resulted in a duplicate-id-aria violation, as domain names appear more often on a page (e.g. court). In order to enable the rule again, an extra “id” property is added to the field configs, in order to decouple id from domain name, so we are able to use different ids in the frontend.