Fix(frontend): remove laziness from text input
The `TextInput` component was using a lazy v-model. As this is a very
basic component, this had a huge effect. Rather than having an opt-out
feature (which isn't fully possible), we decided to make it opt-in. Who
needs laziness, can add this on any layer above.
The original reason for the auto-saving and the document header, do not
have much weight here and do not need to be adapted as agreed.

This allowed the simplification of many tests which always used a tab
key action to leave an input to trigger an update.