Feat(frontend): add conversions to norms service
With the new metadata, more data mapping is necessary in the frontend.
The response and request schema of the backend API do no fully fit the
needs of the frontend with its input components. Therefore some more
complex de- and encoding functions have been added to the norms service.

As we are currently in a bigger transition phase were we skip full
compatibility between a few commits, this tries to be an autonomous
independent commit, but isn't 100%. Some parts might look awkward. But
to avoid a single monster size commit, these packaged smaller commits
are used.

This already uses the new metadata types and section names of the API.

A bunch of data utility functions have been added. While some of them
might be available in a library like Lodash, it is not worth to
introduce such a dependency yet.