Feat(norms): use new metadata on frame route
This does a lot of stuff at once. That's because the whole de- and
encoding was written before and was not made use of. In result a lot of
things have been restructured around this.

Besides using some new input group component for the new metadata
sections, some others got touched to. As the generic `InputGroup` can't
handle our metadata types properly, a lot of typing issues occurred
during the migration. A bunch of input groups have been just a single
input. For those, the single input was just injected into the frame
route component directly.
A bunch of other new metadata values are still mixed with "flat
metadata" values and used in input groups. For these, an intermediate
solution was established.