Test(norms): temporally disable end-to-end tests
The team has decided to work on a separate branch to allow breaking
changes without maintaining the compatibility between frontend and
backend and the functionality of the (end-to-end) tests.
The breaking changes have been finished and the application works
properly again. Though, the end-to-end tests are not yet working again.
To no longer maintain this branch and get back to main, it was decided
to disable the end-to-end tests and get back to main. The argument is
that a separate branch where the end-to-end tests do not run in the
pipeline is very similar to disabled tests on main. Though we have
actual deployment to production.
Anyway we work on the end-to-end tests right away. But it allows us to
work with the whole team on the next story again without the version
control mess.
THIS IS EXCEPTIONAL BEHAVIOR!

This also fixes some typing issues in the test data which are not
reported by our hooks and linting.