Test(norms): fix e2e test to display norm metadata
With the new expandable metadata sections which are repeatable, the
end-to-end tests were not working anymore. Therefore it was necessary
to extend them so they open the sections and analyze every list entry of
it one by one.

To enable the above, two minor changes had to be made in the actual
components to allow addressing them properly.

With this change, a strong type system was introduced for the fixtures
and object used to verify the correct metadata is displayed. This allows
for a much better developer experience as you get code completion, hover
helps and proper linting errors.

Nevertheless the end-to-end tests are still a bit messy and not the best
piece of code. This should be improved step by step.
Though some parts will already get better once we have migrated all
metadata to the new design without this mixture.