Test(norms): shorten e2e norm paragraph text
The new norm we use for the end-to-end tests has lists in its
paragraphs. This is something we can't yet properly verify within the
exported LegalDocML XML file. As we know that this whole topic of the
bill structure will be tackled soon, we do not attempt to solve this
now. We rather decided to "shorten" the tests by not verifying the
complete paragraph, but just the part without the list.