Test(norms): add new set of e2e utilities
Interacting with the inputs for the norm metadata gets more and more
complex. At the same time more tests need to access this functionality.
Therefore a new set of more "generic" utilities have been added to the
end-to-end tests.
The interactions are split into the two types of editing and reading
inputs. For each type of interaction then, multiple levels of
interactions are split into separate functions and composed together.
This also tries to establish a somewhat strong type system around the
input fields.