Test(norms): extend e2e test for new sections
Note that unfortunately we currently fail to properly test the editing
of the new sections. This is an open topic we need to fix, but currently
have not the time to do. A fix will follow soon.
It is also a new challenge to use the new radio buttons within the
sections. Moreover we also need to adapt the usage of dropdowns in
relation to metadata enumerations.