Feat(frontend): allow to hide input errors
It is common sense in the team that the current displaying of errors in
the `InputElement` isn't too nice. Especially as this makes it hard to
use it without an error and have a clean layout. As the error section
has a static height to avoid layout jumping, this looks ugly.

As it is very hard to solve the origin problem and refactor the error
handling, a new flag was added that allows to disable this error
message. This allows to use the `InputElement` with a clean layout.

And as the norms team has no input errors at all right now, this is safe
to use for now.