Fix(frontend): remove margin from expandable
It is in general a bad idea to have margins in re-usable components as
it makes it hard to use them in various layouts.
We agreed in both teams that we don't need it there. Who ever needs
a margin for his layout must do it as usual in the parent component.