Feat(norms): use single field sections
Apply the single field section design on the frame route.

This required the end-to-end tests to change a little as these work
differently than the old input groups and also the data set sections.

The flat metadata of the frame route component could now be strongly
typed again, as no intermediate metadata is merged anymore. Anyway, due
to the new usage, it must change from a computed value with getter and
setter to a reference with watchers.