Test(norms): remove usage of hamcrest assertions
The team decided to use only a single assertion library over the whole
project. We already added a code convention test to verify that we do
not use `JUnit` assertions. Though there was still one test file which
was using `Hamcrest` assertions. These have been replaced with the
`AssertJ` parts and the code convention test was extended.

RISDEV-0000