Feat(frontend): make date input not lazy
The `DateInput` component did only emit model updates when the user
leaves the input. This is an issue for reactive situations as the norms
team has now.
The still remaining limitation is that only valid dates are emitted. As
soon as a date becomes invalid, an undefined value is emitted again.

The tests of this component are quite messy and could be improved in
future.

RISDEV-1217