Make date input not lazy, handle formatting
RISDEV-1442
Before, we only emitted onBlur. Now the emit is being fired, as soon as we have a valid input, which is with a day, a month and at least one character as a year. As the dayjs produced weird outputs with 1 or 2 digit year inputs (eg. 0002 --> 1902), this lib was replaced with a simple javascript Date formatting.