Fixed copy-paste from side panel test
The reason for the test to fail that the test expected the text area to contain a paragraph with a span, containing the left aligned text. However, the paragraph somehow gets a `data-pm-slice` attribute which is not expected in the test: <p data-pm-slice="1 1 []"><span style="color: rgb(0, 0, 0)">I am left aligned</span></p>. The paragraph was therefore removed from the test case.

RISDEV-915