Fix(norms): stable e2e folder creation
During the end-to-end test for norms, a temporary folder gets created to
store test data on download. In theory, the code checks first of the
folder exists and creates the folder only if not. But in scenarios of
high parallelisation on the same machine, it can happen that the folder
does not exist, but another test thread creates the folder in-between,
so the second thread fails to create the folder as it exist by now.

The fix for this problem is to simply get rid of the check at all and
set some make directory command options that take care of this issue.

RISDEV-0000