Chore(frontend): remove async computed setter
We got a reported code smell in the frontend where we use an
asynchronous setter function for a computed value. In general it is
pretty bad practice to do asynchronous stuff in such a setter. Anyway
there have been attempts to resolve it which failed for unknown reasons.
In any case, the code smell can be simply removed by marking the
function as synchronous and not awaiting the promises in the function body.

RISDEV-0000