Feat(norms): improve expandable data set a11y
This improve the accessibility of the expandable data sets. Or rather
the used components around it. So the icon used for the open and closed
state has now a matching aria-label. This is relevant for screen-reader
as they do read the icon name else.
Furthermore it is now possible to access the entries of an editable list
with the keyboard. Although it was always possible to access the two
buttons for each entry to edit and delete it, it was hard to determine
for which entries these buttons are. Having the content of the item
itself as the previous tab index, allows screen-readers to read the
content first.

RISDEV-1209