Test(norms): prevent cross adapter dependencies
The different adapters in the backend for the norms should not depend on
each other. Dependencies must only point inwards or outwards, but never
cross.
In this case, the database output adapter was re-using an encoding function
from the REST controller of an input port.

To prevent such issues, a new code convention/architecture fitness test
has been added which ensures this.

RISDEV-0000