Fix(norms): fix missing ELIs when searching norms
The ELI property of norms returned by the search is always empty. That is because
the majority of information the ELI for a norm depends on has moved into the new
metadata structure. As the search norms output port isn't fully implemented but
was (accidentally?) optimized to load less data, the metadata is missing. Thereby
the ELI and it ends up to be empty.
To solve this issue, the metadata tables must be read and loaded too.

Note that we intend to fix this search norms adapter implementation
anyways. This is just a first quick fix for an urgent problem as the API
is broken by this.

RISDEV-1657