Feat(frontend): improve editable list component
When the editable list includes only a single entry it will
automatically put this into edit mode on initial render. This also
solves the issue after deleting all entries in the list, closing and
opening it again, that there is no more an awkward empty list entry.

This caused a chain reaction, as the interactions with the lists
changed. Especially for the end-to-end tests. The metadata section
utilities had to be extended to recognize and address these cases.
While doing so, the code was split a little more to maintain the
abstraction level of the functions. This hopefully contributes to the
readability of it.

RISDEV-1190