Test(norms): make use of recursive AssertJ
In the unit and integration tests of the norms backend, we introduced
a bunch of custom reflection assert functions. Now we learned more about
our assertion library and figured out how to replace these custom
functions with functionality from the library. Thereby we don't have to
maintain our own functions which were hard to read anyway. The
streamline code of AssertJ is nice to read and powerful.

There is a single last custom assertion function remaining which could
not have been replaced yet as it is much more complicated. Potentially
even questionable if useful. Anyway, this is a first huge improvement.

RISDEV-0000