Fix(norms): filters empty sections on encoding
There were issues where the frontend for the norms was sending empty
metadata sections to the backend. As the backend validated these
sections as invalid, the save operation always failed.
This issue happened for example if the user adds a new official
reference entry, switches the type of the announcement and then inserts
no values. As long as this empty entry gets not removed, an empty
sections will be send.

As this issue comes from the v-model data handling of the editable list
and other components, it is much simpler, less error prone and secure to
filter such sections during the encoding process for the API. Especially
as we already had such a filter. The filter was just not working for all
cases.
A new test case is reproducing the issue and proofs that it was fixed.

RISDEV-1209