Prototype: add static build for footnotes
We started with doing proper prototypes of the frontend for discovery
purpose. There is a need to share these prototypes with the users to
allow them trying it out interactively.
As there is no proper solution for this yet, a first quick-and-dirty
approach is to put it as a sub-route on our GitHub pages route. This
will be removed again and resolved with a proper solution.

RISDEV-1706