Only run a11y tests in a11y-job
The `browser` option of the workflow actually relates to a playwright
project. Since we have a dedicated project for a11y now, we need to set
the browser to `a11y`.

It would be nice to only use projects and not rely on package.json
scripts. However, we need the option to only run the e2e tests of a give
substream with a single browser.

Addresses RISDEV-1590