Stop propagation of key event when deleting chips
We already handle the event, to we should stop its propagation.
Otherwise it might trigger other event handlers (e.g. pressing enter
deletes chips, but also finishes + exists edit mode in editable lists.
So deleting a chip within an item in an editable list also exits the
edit mode of that item, which is unintended).

RISDEV-0000