Test(norms): use specific tags for zip download
During the norms end-to-end tests, we download an example Juris Zip
file. This get fetched from a private GitHub repository due to copyright
issues. Anyway, during the development process of this repository, the
same Zip file gets update/overwritten with new features. Thereby it can
lead to issues, if the tests here just always download the latest
version of that file even if the frontend is not yet compatible with
them. In the past it lead to a failing pipeline/end-to-end tests, if the
Zip was just updated.
To decouple these two, the tests do now download a specific version of
the Zip. As we use Git tags for our versions for the converter
repository, we use them here to download the latest compatible version.

This also solves the issue of a stale cache, especially for developers
rarely restarting their PC (which clear the temporary folder). The Zip
files get now saved into a folder with the tag as name. Thereby always
the correct version is used and re-downloaded as long as the Git tags
remain stable.

RISDEV-1375