Feat(norms): remove flat categorized reference
This cleans up the intermediate state where the categorized norm is
already part of the metadata structure as multi-entry, but we still
support the flat metadata field on a norm. As all system components have
updates to the new metadata structure, the support and everything
related to it can be dropped.

RISDEV-1375