Use mocked routes in route guard tests
This will reduce complexity in the tests because we don't need to
support/mock every single thing routes do. The specific routes are also
not relevant for the tests, they just happend to be there as a side
effect of a module we're importing.

In this specific case, this fixes an issue with newer versions of vitest
not being able to resolve the routes geenrated by the pages plugin.

RISDEV-0000