Feat(norms): chip inputs for announcements
The two metadata fields for "additional info" and "explanation" should
be able to be repeated as multi-entry. This is the case for all
announcement sections.

This is the first case where we have a component test which uses
multiple chip inputs. Currently, the chip input is not testing friendly.
Should mean it is hard to address the chips and the "child" input of it
to access values and interact with it.
While this is an already known and open topic, only a quick fix has been
established for this feature. This allows to directly address the parent
chip input container and also the chip values with test identifier
properties.

Some code changes are not directly related to this feature but try to
improve some code readability at close locations.

RISDEV-1969