Feat(norms): adapt announcement date conversions
The announcement date is now also available in the metadata sections.
For some compatibility concerns, the flat part is still there too. In
result the conversion from Juris XML to the domain, we store the
announcement date into both. The over way around, only the new format is
used now.

RISDEV-1965