Chore(frontend): use new Vue defineEmit syntax
This new syntax was introduced with Vue 3.3 and provides a more succinct
syntax. This syntax should also help to resolve many (false positive)
code-smells at the quality gate.

This is a bulk edit.
Some locations haven't been transformed as the events where defined in
an interface that was deeply used elsewhere too. This is an open topic
anyway as this is mostly the case for the `useInputModel` composable.

RISDEV-0000