Refactor: replace InputElements with components
Removes the intermediary step of using an InputElement which will then
render the actual input component in favor of just rendering the
components directly. This allows us to simplify and unify error handling
and improves DX because we can now have TypeScript warnings and hints on
the components we're using directly.

RISDEV-0000