Prevent saving with validation errors
RISDEV-2164
The prevention of  saving of norms with a single norm validation, makes the code much easier. Before we needed access to the validation error in the input group (edit mode) und in the list view. That's why the validation error was stored in the NormReference class, which caused continuous automatic saving. Because a norm should not be saved with a validation error, we could now simplify the code by removing the validation logic from the class and just validate in the input group.