Chore(norms): break cross dependencies
We have an architecture fitness test that it intended to ensure that no
adapter depends on any other. Turned out this test is not working
properly and we established cross dependencies.
In fact some output adapters (e.g. for the database) use decoding
functions from the input controller for the REST API.

This is an issue. It has been discovered that the decoding needs to
change for the input adapter. But this breaks all the other adapters
which depend on this too. That should not be the case due to proper
independence.

While the test is still broken (unfortunately don't know how to fix it
right now), this breaks this dependency for now.

RISDEV-0000