Fix(norms): metadatum value not nullable generic
Per default, Kotlin its generic types are always nullable. This is a bit
unintuitive, because everywhere else it is the opposite.
The issue is that our Metadatum entity is generic over its value
property. That means also over Null. But our specification that checks
the value type does not allow Null. As this specification is currently
required with the `init` hooks, this leads to runtime issues.

As this is a nullable topic, it is possible to get some compile time
support to enforce no Null usage for Metadatum values. This ensures
during compile time that we do no mistakes. This is the opposite to the
specifications, which are explicitly runtime specific.

Setting the Metadatum to be generic over a not nullable type, uncovered
sections in our code that made use of it like a "feature". This was
especially the case for the import from Juris and the data mapping. The
result was a failed import where the importer should actually just ignore
this Null field.

RISDEV-2153