Refactor: use new emit syntax, simplify v-model
This uses the new syntax for declaring emit events in Vue components
that's available since Vue 3.3 to avoid code smell complaints form
SonarCloud. It also refactors the component to use a computed v-model
instead of the input model composable, as that is not compatible with
the new event syntax.

Also removes the manual declaration and re-emitting of the input event.
This should work out of the box thanks to fall-through attributes. Also
updates the unit tests to ensure this is in fact the case.

RISDEV-0000