Feat(norms): use 2 rows as default in textarea
Having two rows per default should make it more obvious to the user that
he is not in a simple text input field, but an area.

For specific cases it can still be set to 1 row. This was the case for
the official long title. As this input is "hidden" behind the
InputElement. To make it work it was necessary to extend the respective
input attribute and forward them.

RISDEV-2055