Fix(norms): hide scrollbar on autosize textarea
When the TextAreaInput component is set to automatically resize to the
contents height, it still happened in some browsers that a scrollbar
appeared. That shouldn't be the case as the height is actually enough.
To prevent the visual bug of the scrollbar, it gets explicitly disabled
in that case.

RISDEV-2055