Feat(norms): add domain specification result
The specification pattern in the norms domain gets extended. It needs to
be capable of providing informations why a specification is not
satisfied for a certain instance.
In preparation to this change, a new data class `SpecificationResult`
gets introduced. It comes with a few handy features to improve the
usability later and to support good to read code.

RISDEV-1680