Feat(norms): extend specification concept
This extends specifications to use the recently added
`SpecificationResult` as "result" instead of a plain boolean. This
caused some breaking changes.
First of all, the usage has become a two step process. First the
evaluation call and then the inspection of the result.
Second, the compostability has been reduced to only ANDing
specifications. This is because other compositions become very hard to
do when the results are no more simple booleans.

Additionally a bunch of convenience features have been added to make the
usability more comfortable and to enable clean code.

In the end, this change required to adapt the usage of the
specification. Please note that the specification definitions have been
updated in the most simple possible way. Future improvements are
planned.

RISDEV-1680